Move UT of store_to_put to a separated file
